+++ root/src_code/utils/Optional.h
@@ -5,10 +5,10 @@
 #define USING_SDL
 
 namespace AVCallback {
-using VideoPlayer = std::function<void(void*, void*, uint32_t, int, int, int64_t)>;
+using VideoPlayer = std::function<void(void*, void*, uint32_t, int, int, double)>;
 void* initVideoPlayer(const std::string& uid, VideoPlayer* f);
 
-using PcmPlayer = std::function<void(void*, void*, uint32_t, int64_t)>;
+using PcmPlayer = std::function<void(void*, void*, uint32_t, double)>;
 void* initPcmPlayer(const std::string& uid, PcmPlayer* f);
 
 void destroyPcmPlayer(void* handle);
@@ -50,7 +50,7 @@ struct PlayInternal {
         handle = nullptr;
     }
 
-    void Play(void* data, uint32_t size, int64_t pts) {
+    void Play(void* data, uint32_t size, double pts) {
         std::lock_guard<std::mutex> lock(mtx_);
         if (handle) {
             player(handle, data, size, pts);
@@ -81,7 +81,7 @@ struct PlayInternal {
         handle = nullptr;
     }
 
-    void Play(void* data, uint32_t size, int w, int h, int64_t pts) {
+    void Play(void* data, uint32_t size, int w, int h, double pts) {
         std::lock_guard<std::mutex> lock(mtx_);
         if (handle) {
             player(handle, data, size, w, h, pts);
